requirements.txt
matplotlib==3.2.2
numpy==1.17.0
streamlit==0.73.1
Pillow==8.1.0
scikit_learn==0.24.0
